Implementation of Guest Controls
Created a userSettings json file to store each users unique setting preferences (also accomidated for testers with test json file and update in global variable file). When a User logs in, their settings are now initialized in the user global variable, and when a user creates an account, all settings are default and stored in the userSettings json file. The guest control page is made where users can change their settings and is updated to their file. Currently it is not implemented in the InCollege app as Mayank is going to integrate it.